Split feature detector by OS. Build a dummy version on Windows.
